# [router] PD Router Simplification and Reorganization